backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
openai==1.70.0
langchain==0.3.22
supabase==2.15.0
pydantic==2.11.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9 